Moves CryptoAPI and CryptoSDK into its own location rather that hidden in payments.rs.  Also cleans up a ton of unused use statements
