+++ libsovtoken/src/logic/indy_sdk_api/mod.rs
@@ -1,0 +1,3 @@
+//! for all wrappers/facades/etc around rust-indy-sdk crate or indy-sdk itself
+
+pub mod crypto_api;
